don't show expanded forms
As the list of experiments grows, this section becomes more complex
to maintain.  It will also get more complex as experiments end,
either in success (warning removed) or failure (feature not
supported, enabling should throw an error).

We could generate this Pod later if we want, but for the most part,
I think it's much simpler to just omit this section for now.